Index

AngularPowerSpectra.channelindex

AngularPowerSpectra.decouple_covmat

AngularPowerSpectra.kᵤ

AngularPowerSpectra.map2cl

AngularPowerSpectra.mask!

AngularPowerSpectra.mcm

AngularPowerSpectra.quickpolΞ!

AngularPowerSpectra.quickpolΞ!

AngularPowerSpectra.synalm

AngularPowerSpectra.synalm!

AngularPowerSpectra.Ξsum

AngularPowerSpectra.channelindex — Method

channelindex(s)

Convert string/char T,E,B => 1,2,3

Examples

julia> channelindex("E")
2

AngularPowerSpectra.decouple_covmat — Method

decouple_covmat(Y, B1, B2; lmin1=2, lmin2=2)

Decouples a covariance matrix Y, performing B₁⁻¹ × M × (B₂⁻¹)^† by mutating M. Zeros out ℓ₁, ℓ₂ within BOTH lmin1 and lmin2 (i.e the corner).

AngularPowerSpectra.kᵤ — Method

kᵤ([T=Float64], u)

Defined only for u ∈ {-2, 0, 2}.

AngularPowerSpectra.map2cl — Method

map2cl(...)

Arguments:

map_1::Map{T}: masked map

map_2::Map{T}: masked map

factorized_mcm::Factorization: lu(mode coupling matrix)

Bℓ_1::SpectralVector{T}: beam associated with first map

Bℓ_2::SpectralVector{T}: beam associated with second map

Returns:

Array{T,1}: spectrum

AngularPowerSpectra.mask! — Method

mask!

Convenience function for applying a mask to a map.

AngularPowerSpectra.mcm — Method

mcm(workspace::SpectralWorkspace{T}, spec::MapType, f1_name::String, f2_name::String) where {T}

Arguments:

workspace::SpectralWorkspace{T}: stores the SHTs of the masks

spec::String: the spectrum to compute, such as "TT", "TE", or "EE"

f1_name::String: the name of the first field

f2_name::String: the name of the second field

Returns:

SpectralArray{T,2}: zero-indexed array containing the mode-coupling matrix

Examples

m1 = PolarizedField("field1", mask1_T, mask1_P)
m2 = PolarizedField("field2", mask2_T, mask2_P)
workspace = SpectralWorkspace(m1, m2)
𝐌 = mcm(workspace, spec, "field1", "field2")

AngularPowerSpectra.quickpolΞ! — Method

quickpolΞ!(𝚵::AA, ν₁, ν₂, u₁, u₂, s₁, s₂, ω₁, ω₂, b₁, b₂, buf1, buf2)

This computes the $\Xi_{\ell^{\prime \prime},\ell}$ matrix. It assumes $\rho$ has been absorbed into the $\omega$ terms.

ω₁: effective scan weights with spin s₁ + ν₁

ω₂: effective scan weights with spin s₂ + ν₂

b₁: inverse noise-weighted beam multipoles for spin u₁, detector j₁

b₂: inverse noise-weighted beam multipoles for spin u₂, detector j₂

AngularPowerSpectra.synalm! — Method

synalm!([rng=GLOBAL_RNG], Cl::AbstractArray{T,3}, alms::Vector{Alm{Complex{T}}}) where T

In-place synthesis of spherical harmonic coefficients, given spectra.

Arguments:

Cl::AbstractArray{T,3}: array with dimensions of comp, comp, ℓ

alms::Vector: array of Alm to fill

Examples

nside = 16
C0 = [3.  2.;  2.  5.]
Cl = repeat(C0, 1, 1, 3nside)  # spectra constant with ℓ
alms = [Alm{Complex{Float64}}(3nside-1, 3nside-1) for i in 1:2]
synalm!(Cl, alms)

AngularPowerSpectra.synalm — Method

synalm([rng=GLOBAL_RNG], Cl::AbstractArray{T,3}, nside::Int) where T

Arguments:

Cl::AbstractArray{T,3}: array with dimensions of comp, comp, ℓ

nside::Int: healpix resolution

Returns:

Vector{Alm{T}}: spherical harmonics realizations for each component

Examples

nside = 16
C0 = [3.  2.;  2.  5.]
Cl = repeat(C0, 1, 1, 3nside)  # spectra constant with ℓ
alms = synalm(Cl, nside)

AngularPowerSpectra.Ξsum — Method

Ξsum(alm₁, alm₂, w3j₁, w3j₂)

Sum over $\ell$ and $m$ of two $a_{\ell m}$ and nontrivial Wigner-3j vectors. This is a step in computing the $\mathbf{\Xi}$ matrix. The $\rho$ factors are not in this summation, as they can be pulled out.

\[\begin{aligned} (\Xi \mathrm{sum}) &= \sum_{\ell^{\prime} m^{\prime}} \, \left(_{s_1+\nu_1}\tilde{\omega}^{(j_1)}_{\ell^\prime m^\prime}\right) \left(_{s_2+\nu_2}\tilde{\omega}^{(j_2)}_{\ell^\prime m^\prime}\right)^* \\ &\qquad\qquad \times \begin{pmatrix} \ell & \ell^{\prime} & \ell^{\prime\prime} \\ -s_1 & s_1+\nu_1 & -\nu_1 \end{pmatrix} \begin{pmatrix} \ell & \ell^{\prime} & \ell^{\prime\prime} \\ -s_2 & s_2+\nu_2 & -\nu_2 \end{pmatrix} \end{aligned}\]