Index

AngularPowerSpectra.CovField

AngularPowerSpectra.CovarianceWorkspace

AngularPowerSpectra.CovarianceWorkspace

AngularPowerSpectra.SpectralArray

AngularPowerSpectra.channelindex

AngularPowerSpectra.coupledcov

AngularPowerSpectra.coupledcov

AngularPowerSpectra.decouple_covmat

AngularPowerSpectra.kᵤ

AngularPowerSpectra.kᵤ

AngularPowerSpectra.map2cl

AngularPowerSpectra.mask!

AngularPowerSpectra.mcm

AngularPowerSpectra.mcm

AngularPowerSpectra.quickpolW

AngularPowerSpectra.quickpolW

AngularPowerSpectra.quickpolΞ!

AngularPowerSpectra.quickpolΞ!

AngularPowerSpectra.synalm

AngularPowerSpectra.synalm!

AngularPowerSpectra.Ξsum

AngularPowerSpectra.CovarianceWorkspace — Method

CovarianceWorkspace(m_i, m_j, m_p, m_q; lmax::Int=0)

Inputs and cache for covariance calculations. A covariance matrix relates the masks of four fields and spins. This structure caches various cross-spectra between masks and noise-weighted masks.

Arguments:

m_i::CovField{T}: map i

m_j::CovField{T}: map j

m_p::CovField{T}: map p

m_q::CovField{T}: map q

Keywords

lmax::Int=0: maximum multipole to compute covariance matrix

AngularPowerSpectra.SpectralArray — Method

Zero-based array for representing spectra matrices.

AngularPowerSpectra.channelindex — Method

channelindex(s)

Convert string/char T,E,B => 1,2,3

Examples

julia> channelindex("E")
2

AngularPowerSpectra.coupledcov — Method

coupledcov(ch1, ch2, workspace, spectra;
           noise_ratios=Dict(), lmax=0) where T

Arguments:

ch1::Symbol: spectrum type of first spectrum (i.e. :TT, :TE, :EE)

ch2::Symbol: spectrum type of second spectrum (i.e. :TT, :TE, :EE)

workspace: cache for working with covariances

spectra: signal spectra

Keywords

noise_ratios::AbstractDict: ratio of noise spectra to white noise

lmax=0: maximum multipole moment for covariance matrix

Returns:

SpectralArray{T,2}: covariance matrix (0-indexed)

AngularPowerSpectra.decouple_covmat — Method

decouple_covmat(Y, B1, B2; lmin1=2, lmin2=2)

Decouples a covariance matrix Y, performing B₁⁻¹ × M × (B₂⁻¹)^† by mutating M. Zeros out ℓ₁, ℓ₂ within BOTH lmin1 and lmin2 (i.e the corner).

AngularPowerSpectra.kᵤ — Method

kᵤ([T=Float64], u)

Defined only for u ∈ {-2, 0, 2}.

AngularPowerSpectra.map2cl — Method

map2cl(...)

Arguments:

map_1::Map{T}: masked map

map_2::Map{T}: masked map

factorized_mcm::Factorization: lu(mode coupling matrix)

Bℓ_1::SpectralVector{T}: beam associated with first map

Bℓ_2::SpectralVector{T}: beam associated with second map

Returns:

Array{T,1}: spectrum

AngularPowerSpectra.mask! — Method

mask!(m::Map{T}, mask::Map{T}) where T

Convenience function for applying a mask to a map.

AngularPowerSpectra.mcm — Method

mcm(spec::Symbol, alm₁::Alm{T}, alm₂::Alm{T}; lmax=nothing)

Arguments:

spec::Symbol: cross-spectrum, i.e. :TE

alm₁::Alm{T}: first mask's spherical harmonic coefficients

alm₂::Alm{T}: second mask's spherical harmonic coefficients

Keywords

lmax=nothing: maximum multipole for mode-coupling matrix

Returns:

SpectralArray{T,2}: the index where val is located in the array

AngularPowerSpectra.quickpolW — Method

quickpolW(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}})

Computes a scaled spectrum of the scan pattern.

\[W_{\ell'}^{\nu_1,\nu_2,s_1,s_2,j_1,j_2} = \sum_{m^\prime=-\ell^\prime}^{\ell^\prime} \left(_{s_1+\nu_1}\tilde{\omega}^{(j_1)}_{\ell^\prime m^\prime}\right) \left(_{s_2+\nu_2}\tilde{\omega}^{(j_2)}_{\ell^\prime m^\prime}\right)^*\]

AngularPowerSpectra.quickpolΞ! — Method

quickpolΞ!(𝚵::AA, ν₁, ν₂, s₁, s₂, ω₁, ω₂)

This computes the $\Xi_{\ell^{\prime \prime},\ell}$ matrix. It assumes $\rho$ has been absorbed into the $\omega$ terms.

ω₁: effective scan weights with spin s₁ + ν₁

ω₂: effective scan weights with spin s₂ + ν₂

AngularPowerSpectra.synalm! — Method

synalm!([rng=GLOBAL_RNG], Cl::AbstractArray{T,3}, alms::Vector{Alm{Complex{T}}}) where T

In-place synthesis of spherical harmonic coefficients, given spectra.

Arguments:

Cl::AbstractArray{T,3}: array with dimensions of comp, comp, ℓ

alms::Vector: array of Alm to fill

Examples

nside = 16
C0 = [3.  2.;  2.  5.]
Cl = repeat(C0, 1, 1, 3nside)  # spectra constant with ℓ
alms = [Alm{Complex{Float64}}(3nside-1, 3nside-1) for i in 1:2]
synalm!(Cl, alms)

AngularPowerSpectra.synalm — Method

synalm([rng=GLOBAL_RNG], Cl::AbstractArray{T,3}, nside::Int) where T

Arguments:

Cl::AbstractArray{T,3}: array with dimensions of comp, comp, ℓ

nside::Int: healpix resolution

Returns:

Vector{Alm{T}}: spherical harmonics realizations for each component

Examples

nside = 16
C0 = [3.  2.;  2.  5.]
Cl = repeat(C0, 1, 1, 3nside)  # spectra constant with ℓ
alms = synalm(Cl, nside)

AngularPowerSpectra.Ξsum — Method

Ξsum(alm₁, alm₂, w3j₁, w3j₂)

Sum over $\ell$ and $m$ of two $a_{\ell m}$ and nontrivial Wigner-3j vectors. This is a step in computing the $\mathbf{\Xi}$ matrix. The $\rho$ factors are not in this summation, as they can be pulled out.

\[\begin{aligned} (\Xi \mathrm{sum}) &= \sum_{\ell^{\prime} m^{\prime}} \, W_{\ell'}^{\nu_1,\nu_2,s_1,s_2,j_1,j_2} \times \begin{pmatrix} \ell & \ell^{\prime} & \ell^{\prime\prime} \\ -s_1 & s_1+\nu_1 & -\nu_1 \end{pmatrix} \begin{pmatrix} \ell & \ell^{\prime} & \ell^{\prime\prime} \\ -s_2 & s_2+\nu_2 & -\nu_2 \end{pmatrix} \end{aligned}\]